feat: Implement Faucet and Deposit features
Adds a new Faucet view for obtaining test tokens and introduces a Deposit modal for users to add funds. This also includes updates to the header component to include navigation links for these new features and localization adjustments for the new functionalities.

The Faucet view provides users with test tokens for market exploration and requires BNB for gas fees. The Deposit modal allows for seamless fund deposits and withdrawals, with functionality for copying contract addresses and displaying QR codes.